Use full 36-bit physical address space on SS10
git-svn-id: svn://coreboot.org/openbios/openbios-devel@149 f158a5a8-5612-0410-a976-696ce0be7e32

+++ arch/sparc32/openbios.h
@@ -22,7 +22,7 @@ int openbios(void);
 /* console.c */
 extern void	cls(void);
 #ifdef CONFIG_DEBUG_CONSOLE
-extern int	uart_init(int port, unsigned long speed);
+extern int	uart_init(uint64_t port, unsigned long speed);
 extern void     video_init(void);
 #endif
